add BlogController

+++ project/libs/blog/blog-post/src/blog-post-module/blog-post.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { CreatePostDto } from './dto/create-post.dto';
 import { UpdatePostDto } from './dto/update-post.dto';
 import { BlogPostRepository } from './blog-post.repository';
@@ -15,7 +15,7 @@ export class BlogPostService {
   ) { }
 
   public async create(dto: CreatePostDto): Promise<BlogPostEntity> {
-    const newPost = new BlogPostEntity(dto)
+    const newPost = new BlogPostEntity({...dto, comments: [], likes: []})
     await this.blogPostRepository.save(newPost);
     //await this.notifyService.createPostMail({ id: newPost.id, postDate: newPost.postDate, type: newPost.type, userId: newPost.userId });
     return newPost;
@@ -33,7 +33,11 @@ export class BlogPostService {
     return `This action updates a #${id} blogPostModule`;
   }
 
-  public async remove(id: string): Promise<void> {
+  public async remove(id: string, userId: string): Promise<void> {
+    const post = await this.findById(id);
+    if (post.userId !== userId) {
+      throw new HttpException('Запрещено удалять чужие посты', HttpStatus.BAD_REQUEST);
+    }
     return await this.blogPostRepository.deleteById(id);
   }
